requirements.txt
aiosqlite>=0.19.0
httpx>=0.27.0
pyyaml>=6.0
groq>=0.9.0
openai>=1.40.0
python-telegram-bot>=21.0
fastapi>=0.110.0
uvicorn>=0.30.0
mcp>=1.0.0
cryptography>=42.0
python-dotenv>=1.0.0
rich>=13.0
psutil>=5.9.0
